Revert "accounts-email-token: save/restore URL in email login flow"
